ponderada1/requirements.txt
paho-mqtt   